requirements.txt
python-docx
pandas
scikit-learn
flask
pdfplumber
ollama
